requirements.txt
pillow
ezdxf
tk
